requirements.txt
dspy-ai>=2.0.0
datasets>=2.0.0

